[#258] Extracted BigPipe handling into 'BigPipeTrait'.
